+++ rmk/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2] - 2024-07-12
+
 - Add keyboard macro support
 - Support vial keymap reset command
 - Fix default `lt!` and `lm!` implementation